refactor(button): simplify button imports and enhance dark mode handling
