+++ src/shared/components/layout/header/header.component.ts
@@ -1,9 +1,7 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { SectionTitleComponent } from '../../ui/title/title.component';
-import { LogoComponent } from '../../ui/logo/logo.component';
-import { IconSettingsComponent } from '../../ui/icons/icon-settings/icon-settings.component';
 import { RouterLink } from '@angular/router';
-import { InputComponent } from '@shared/components/ui/input/input.component';
+import { InputComponent, LogoComponent, SectionTitleComponent } from '@shared/components/ui';
+import { IconSettingsComponent } from '@shared/components/ui/icons';
 
 @Component({
   selector: 'app-header',
