+++ src/shared/components/ui/index.ts
@@ -1,2 +1,5 @@
 export * from './logo/logo.component';
 export * from './title/title.component';
+export * from './button/button.component';
+export * from './input/input.component';
+export * from './modal/modal.component';
