+++ src/shared/interfaces/index.ts
@@ -1,2 +1,3 @@
 export * from './ui/link.interface';
 export * from './ui/settings-selection.interface';
+export * from './ui/modal.interface';
